+++ packages/sheets-plugin-numfmt/src/commands/operators/close.numfmt.panel.operator.ts
@@ -1,13 +1,9 @@
-import { ISidebarService } from '@univerjs/base-ui';
 import { CommandType, ICommand } from '@univerjs/core';
-import { IAccessor } from '@wendellhu/redi';
 
 export const CloseNumfmtPanelOperator: ICommand = {
     id: 'sheet.operation.close.numfmt.panel',
     type: CommandType.OPERATION,
-    handler: async (accessor: IAccessor) => {
-        const sidebarService = accessor.get(ISidebarService);
-        sidebarService.close();
-        return true;
-    },
+    handler: () =>
+        // do nothing,just notify panel is closed
+        true,
 };
