+++ packages/sheets-plugin-numfmt/src/components/more-numfmt-type/MoreNumfmtType.tsx
@@ -1,0 +1,15 @@
+import './index.less';
+
+import { MoreDownSingle } from '@univerjs/icons';
+
+export const MoreNumfmtType = (props: any) => {
+    const value = props.value ?? '常规';
+    return (
+        <span className="more-numfmt-type">
+            {value}
+            <div className="icon">
+                <MoreDownSingle />
+            </div>
+        </span>
+    );
+};
