+++ packages/sheets-plugin-numfmt/src/components/more-numfmt-type/index.less
@@ -1,0 +1,15 @@
+.more-numfmt-type {
+    color: var(--black);
+    font-size: 13px;
+    padding: 0px 4px;
+    display: flex;
+    .icon{
+        display: flex;
+        height: auto;
+        font-size: 10px;
+        color: rgb(var(--grey-400));
+        align-items: center;
+        padding-left: 8px;
+
+    }
+}
