+++ packages/sheets-plugin-numfmt/src/controllers/numfmt.controller.ts
@@ -1,0 +1,701 @@
+import numfmt from '@univerjs/base-numfmt-engine';
+import { IRenderManagerService } from '@univerjs/base-render';
+import {
+    EffectRefRangeParams,
+    EffectRefRangId,
+    handleDeleteRangeMoveLeft,
+    handleDeleteRangeMoveUp,
+    handleInsertCol,
+    handleInsertRangeMoveDown,
+    handleInsertRangeMoveRight,
+    handleInsertRow,
+    handleIRemoveCol,
+    handleIRemoveRow,
+    handleMoveRange,
+    ISelectionWithStyle,
+    RefRangeService,
+    runRefRangeMutations,
+    SelectionManagerService,
+    SetRangeValuesCommand,
+} from '@univerjs/base-sheets';
+import { ComponentManager, IMenuService, ISidebarService } from '@univerjs/base-ui';
+import {
+    CellValueType,
+    Disposable,
+    DisposableCollection,
+    ICommandService,
+    INTERCEPTOR_POINT,
+    IUniverInstanceService,
+    LifecycleStages,
+    LocaleService,
+    LocaleType,
+    Nullable,
+    OnLifecycle,
+    Range,
+    SheetInterceptorService,
+    ThemeService,
+    toDisposable,
+} from '@univerjs/core';
+import { SheetSkeletonManagerService } from '@univerjs/ui-plugin-sheets';
+import { IDisposable, Inject, Injector } from '@wendellhu/redi';
+import { Observable } from 'rxjs';
+import { bufferTime, distinctUntilChanged, filter, map, switchMap } from 'rxjs/operators';
+
+import { SHEET_NUMFMT_PLUGIN } from '../base/const/PLUGIN_NAME';
+import { NumfmtItem } from '../base/types';
+import { AddDecimalCommand } from '../commands/commands/add.decimal.command';
+import { SetCurrencyCommand } from '../commands/commands/set.currency.command';
+import { SetNumfmtCommand, SetNumfmtCommandParams } from '../commands/commands/set.numfmt.command';
+import { SubtractDecimalCommand } from '../commands/commands/subtract.decimal.command';
+import {
+    factorySetNumfmtUndoMutation,
+    SetNumfmtMutation,
+    SetNumfmtMutationParams,
+} from '../commands/mutations/set.numfmt.mutation';
+import { CloseNumfmtPanelOperator } from '../commands/operators/close.numfmt.panel.operator';
+import { OpenNumfmtPanelOperator } from '../commands/operators/open.numfmt.panel.operator';
+import { SheetNumfmtPanel, SheetNumfmtPanelProps } from '../components/index';
+import { zhCn } from '../locale/zh-CN';
+import { AddDecimalMenuItem, CurrencyMenuItem, FactoryOtherMenuItem, SubtractDecimalMenuItem } from '../menu/menu';
+import { NumfmtService } from '../service/numfmt.service';
+import { getPatternPreview, getPatternType } from '../utils/pattern';
+
+const createCollectEffectMutation = () => {
+    type Config = { workbookId: string; worksheetId: string; row: number; col: number; value: Nullable<NumfmtItem> };
+    let list: Config[] = [];
+    const add = (workbookId: string, worksheetId: string, row: number, col: number, value: Nullable<NumfmtItem>) =>
+        list.push({ workbookId, worksheetId, row, col, value });
+    const getEffects = () => list;
+    const clean = () => {
+        list = [];
+    };
+    return {
+        add,
+        getEffects,
+        clean,
+    };
+};
+@OnLifecycle(LifecycleStages.Ready, NumfmtController)
+export class NumfmtController extends Disposable {
+    // collect effect mutations when edit end and push this to  commands stack in next commands progress
+    private _collectEffectMutation = createCollectEffectMutation();
+    private _previewPattern = '';
+
+    constructor(
+        @Inject(SheetInterceptorService) private _sheetInterceptorService: SheetInterceptorService,
+        @Inject(ThemeService) private _themeService: ThemeService,
+        @Inject(Injector) private _injector: Injector,
+        @Inject(IUniverInstanceService) private _univerInstanceService: IUniverInstanceService,
+        @Inject(RefRangeService) private _refRangeService: RefRangeService,
+        @Inject(SheetSkeletonManagerService) private _sheetSkeletonManagerService: SheetSkeletonManagerService,
+        @Inject(ICommandService) private _commandService: ICommandService,
+        @Inject(SelectionManagerService) private _selectionManagerService: SelectionManagerService,
+        @Inject(IRenderManagerService) private _renderManagerService: IRenderManagerService,
+        @Inject(NumfmtService) private _numfmtService: NumfmtService,
+        @Inject(ComponentManager) private _componentManager: ComponentManager,
+        @Inject(IMenuService) private _menuService: IMenuService,
+        @Inject(ISidebarService) private _sidebarService: ISidebarService,
+        @Inject(LocaleService) private _localeService: LocaleService
+    ) {
+        super();
+        this._initInterceptorCellContent();
+        this._initInterceptorEditorStart();
+        this._initInterceptorEditorEnd();
+        this._initInterceptorCommands();
+        this._registerRefRange();
+        this._initRealTimeRenderingInterceptor();
+        this._initMenu();
+        this._initLocal();
+        this.initPanel();
+        this._initCommands();
+    }
+
+    openPanel = () => {
+        const sidebarService = this._sidebarService;
+        const selectionManagerService = this._selectionManagerService;
+        const commandService = this._commandService;
+        const univerInstanceService = this._univerInstanceService;
+        const numfmtService = this._numfmtService;
+        const localeService = this._localeService;
+
+        const selections = selectionManagerService.getSelectionRanges() || [];
+        const range = selections[0];
+
+        if (!range) {
+            return false;
+        }
+        const workbook = univerInstanceService.getCurrentUniverSheetInstance();
+        const sheet = workbook.getActiveSheet();
+
+        const cellValue = sheet.getCellRaw(range.startRow, range.startColumn);
+        const numfmtValue = numfmtService.getValue(
+            workbook.getUnitId(),
+            sheet.getSheetId(),
+            range.startRow,
+            range.startColumn
+        );
+        let pattern = '';
+        if (numfmtValue) {
+            pattern = numfmtValue.pattern;
+        }
+        // eslint-disable-next-line no-magic-numbers
+        const defaultValue = (cellValue?.t === CellValueType.NUMBER ? cellValue.v : 12345678) as number;
+
+        const props: SheetNumfmtPanelProps = {
+            onChange: (config) => {
+                if (config.type === 'change') {
+                    this._previewPattern = config.value;
+                    this._renderManagerService.getRenderById(workbook.getUnitId())?.mainComponent?.makeDirty();
+                } else if (config.type === 'confirm') {
+                    const selections = selectionManagerService.getSelectionRanges() || [];
+                    const params: SetNumfmtCommandParams = { values: [] };
+                    const patternType = getPatternType(config.value);
+                    selections.forEach((rangeInfo) => {
+                        Range.foreach(rangeInfo, (row, col) => {
+                            params.values.push({
+                                row,
+                                col,
+                                pattern: config.value,
+                                type: patternType,
+                            });
+                        });
+                    });
+                    commandService.executeCommand(SetNumfmtCommand.id, params);
+                    sidebarService.close();
+                } else if (config.type === 'cancel') {
+                    sidebarService.close();
+                }
+            },
+            value: { defaultPattern: pattern, defaultValue },
+        };
+
+        sidebarService.open({
+            header: { title: localeService.t('sheet.numfmt.title') },
+            children: {
+                label: SHEET_NUMFMT_PLUGIN,
+                ...(props as any),
+            },
+            onClose: () => {
+                commandService.executeCommand(CloseNumfmtPanelOperator.id);
+            },
+        });
+    };
+
+    private _initCommands() {
+        [
+            AddDecimalCommand,
+            SubtractDecimalCommand,
+            SetCurrencyCommand,
+            OpenNumfmtPanelOperator,
+            CloseNumfmtPanelOperator,
+            SetNumfmtCommand,
+        ].forEach((config) => this._commandService.registerCommand(config));
+    }
+
+    private initPanel() {
+        this._componentManager.register(SHEET_NUMFMT_PLUGIN, SheetNumfmtPanel);
+    }
+    private _initLocal = () => {
+        this._localeService.load({ [LocaleType.ZH_CN]: zhCn, [LocaleType.EN_US]: zhCn });
+    };
+
+    private _initMenu() {
+        [AddDecimalMenuItem, SubtractDecimalMenuItem, CurrencyMenuItem, FactoryOtherMenuItem]
+            .map((factory) => factory(this._componentManager))
+            .forEach((configFactory) => {
+                this._menuService.addMenuItem(configFactory(this._injector));
+            });
+    }
+
+    private _initInterceptorCommands() {
+        const self = this;
+        this._sheetInterceptorService.interceptCommand({
+            getMutations(command) {
+                const workbookId = self._univerInstanceService.getCurrentUniverSheetInstance().getUnitId();
+                const worksheetId = self._univerInstanceService
+                    .getCurrentUniverSheetInstance()
+                    .getActiveSheet()
+                    .getSheetId();
+                switch (command.id) {
+                    case SetRangeValuesCommand.id: {
+                        const list = self._collectEffectMutation.getEffects();
+                        if (!list.length) {
+                            return {
+                                redos: [],
+                                undos: [],
+                            };
+                        }
+                        const redos: SetNumfmtMutationParams = {
+                            workbookId,
+                            worksheetId,
+                            values: list.map((item) => ({
+                                pattern: item.value ? item.value.pattern : '',
+                                row: item.row,
+                                col: item.col,
+                                type: item.value?.type,
+                            })),
+                        };
+                        const undos = factorySetNumfmtUndoMutation(self._injector, redos);
+                        return {
+                            redos: [{ id: SetNumfmtMutation.id, params: redos }],
+                            undos: [{ id: SetNumfmtMutation.id, params: undos }],
+                        };
+                    }
+                }
+                return {
+                    redos: [],
+                    undos: [],
+                };
+            },
+        });
+    }
+
+    private _initInterceptorCellContent() {
+        this._sheetInterceptorService.intercept(INTERCEPTOR_POINT.CELL_CONTENT, {
+            handler: (cell, location, next) => {
+                const workbookId = location.workbookId;
+                const sheetId = location.worksheetId;
+                const numfmtValue = this._numfmtService.getValue(workbookId, sheetId, location.row, location.col);
+                if (!numfmtValue) {
+                    return next(cell);
+                }
+                const originCellValue = location.worksheet.getCellRaw(location.row, location.col);
+                if (!originCellValue) {
+                    return next(cell);
+                }
+                // just handle number
+                if (originCellValue.t !== CellValueType.NUMBER) {
+                    return next(cell);
+                }
+
+                let numfmtRes: string = '';
+                if (numfmtValue._cache && numfmtValue._cache.parameters === originCellValue.v) {
+                    return numfmtValue._cache.result;
+                }
+
+                const info = getPatternPreview(numfmtValue.pattern, Number(originCellValue.v));
+
+                numfmtRes = info.result;
+
+                if (!numfmtRes) {
+                    return next(cell);
+                }
+
+                const res = { ...cell, v: numfmtRes };
+
+                if (info.color) {
+                    const color = this._themeService.getCurrentTheme()[`${info.color}500`];
+
+                    if (color) {
+                        res.s = { cl: { rgb: color } };
+                    }
+                }
+
+                numfmtValue._cache = {
+                    result: res,
+                    parameters: originCellValue.v as number,
+                };
+
+                return res;
+            },
+        });
+    }
+
+    private _initRealTimeRenderingInterceptor() {
+        const isPanelOpenObserver = new Observable<boolean>((subscribe) => {
+            this._commandService.onCommandExecuted((commandInfo) => {
+                if (commandInfo.id === OpenNumfmtPanelOperator.id) {
+                    subscribe.next(true);
+                }
+                if (commandInfo.id === CloseNumfmtPanelOperator.id) {
+                    subscribe.next(false);
+                }
+            });
+        });
+        const selectionChangeWithPanelOpen$ = isPanelOpenObserver.pipe(
+            switchMap(
+                (isOpen) =>
+                    new Observable<ISelectionWithStyle[]>((subscribe) => {
+                        const subscription =
+                            isOpen &&
+                            this._selectionManagerService.selectionInfo$.subscribe((selectionInfos) => {
+                                isOpen && selectionInfos && subscribe.next(selectionInfos);
+                            });
+                        return () => subscription && subscription.unsubscribe();
+                    })
+            ),
+            switchMap((selectionInfos) => {
+                const selectionRanges = selectionInfos.map((selectionInfo) => selectionInfo.range);
+                return new Observable<{
+                    disposableCollection: DisposableCollection;
+                    selectionRanges: Array<ISelectionWithStyle['range']>;
+                }>((subscribe) => {
+                    const disposableCollection = new DisposableCollection();
+                    subscribe.next({ disposableCollection, selectionRanges });
+                    return () => {
+                        disposableCollection.dispose();
+                    };
+                });
+            })
+        );
+
+        selectionChangeWithPanelOpen$.subscribe(({ disposableCollection, selectionRanges }) => {
+            this.openPanel();
+            disposableCollection.add(
+                this._sheetInterceptorService.intercept(INTERCEPTOR_POINT.CELL_CONTENT, {
+                    priority: 99,
+                    handler: (cell, location, next) => {
+                        const { row, col } = location;
+                        if (
+                            selectionRanges.find(
+                                (range) =>
+                                    range.startColumn <= col &&
+                                    range.endColumn >= col &&
+                                    range.startRow <= row &&
+                                    range.endRow >= row
+                            )
+                        ) {
+                            const rawValue = location.worksheet.getCellRaw(row, col);
+                            const value = rawValue?.v;
+                            const type = rawValue?.t;
+                            if (value === undefined || value === null || type !== CellValueType.NUMBER) {
+                                return next();
+                            }
+                            // const pattern = ''; // todo get
+                            const info = getPatternPreview(this._previewPattern, value as number);
+                            if (info.color) {
+                                const colorMap = this._themeService.getCurrentTheme();
+                                const color = colorMap[`${info.color}500`];
+                                return {
+                                    ...cell,
+                                    v: info.result,
+                                    t: CellValueType.STRING,
+                                    s: { cl: { rgb: color } },
+                                };
+                            }
+                            return {
+                                ...cell,
+                                v: info.result,
+                                t: CellValueType.STRING,
+                            };
+                        }
+                        return next();
+                    },
+                })
+            );
+        });
+    }
+
+    /**
+     * Process the initial values before entering the edit
+     * @private
+     * @memberof NumfmtService
+     */
+    private _initInterceptorEditorStart() {
+        this._sheetInterceptorService.intercept(INTERCEPTOR_POINT.BEFORE_CELL_EDIT, {
+            handler: (value, context, next) => {
+                const row = context.row;
+                const col = context.col;
+                const numfmtCell = this._numfmtService.getValue(context.workbookId, context.worksheetId, row, col);
+                if (numfmtCell) {
+                    const type = getPatternType(numfmtCell.pattern);
+                    switch (type) {
+                        case 'scientific':
+                        case 'percent':
+                        case 'currency':
+                        case 'grouped':
+                        case 'number': {
+                            // remove the style atr
+                            const cell = context.worksheet.getCellRaw(row, col);
+                            return cell ? filterAtr(cell, ['s']) : cell;
+                        }
+                        case 'date':
+                        case 'time':
+                        case 'datetime':
+                        default: {
+                            return next && next(value);
+                        }
+                    }
+                }
+                return next(value);
+            },
+        });
+    }
+
+    /**
+     * Process the  values after  edit
+     * @private
+     * @memberof NumfmtService
+     */
+    private _initInterceptorEditorEnd() {
+        this._sheetInterceptorService.intercept(INTERCEPTOR_POINT.AFTER_CELL_EDIT, {
+            handler: (value, context, next) => {
+                // clear the effect
+                this._collectEffectMutation.clean();
+                const currentNumfmtValue = this._numfmtService.getValue(
+                    context.workbookId,
+                    context.worksheetId,
+                    context.row,
+                    context.col
+                );
+                const clean = () => {
+                    currentNumfmtValue &&
+                        this._collectEffectMutation.add(
+                            context.workbookId,
+                            context.worksheetId,
+                            context.row,
+                            context.col,
+                            null
+                        );
+                };
+                if (!value?.v) {
+                    clean();
+                    return next(value);
+                }
+
+                const content = String(value.v);
+
+                const dateInfo = numfmt.parseDate(content) || numfmt.parseTime(content);
+                const isTranslateDate = !!dateInfo;
+                if (isTranslateDate) {
+                    if (dateInfo && dateInfo.z) {
+                        const v = Number(dateInfo.v);
+                        this._collectEffectMutation.add(
+                            context.workbookId,
+                            context.worksheetId,
+                            context.row,
+                            context.col,
+                            {
+                                type: 'date',
+                                pattern: dateInfo.z,
+                            }
+                        );
+                        return { ...value, v, t: CellValueType.NUMBER };
+                    }
+                } else if (
+                    ['date', 'time', 'datetime'].includes(currentNumfmtValue?.type || '') ||
+                    !isNumeric(content)
+                ) {
+                    clean();
+                }
+                return next(value);
+            },
+        });
+    }
+
+    private _registerRefRange() {
+        this.disposeWithMe(
+            toDisposable(
+                this._sheetSkeletonManagerService.currentSkeleton$
+                    .pipe(
+                        map((skeleton) => skeleton?.sheetId),
+                        distinctUntilChanged(),
+                        switchMap(
+                            () =>
+                                new Observable<DisposableCollection>((subscribe) => {
+                                    const disposableCollection = new DisposableCollection();
+                                    subscribe.next(disposableCollection);
+                                    return () => {
+                                        disposableCollection.dispose();
+                                    };
+                                })
+                        )
+                    )
+                    .subscribe((disposableCollection) => {
+                        const workbook = this._univerInstanceService.getCurrentUniverSheetInstance();
+                        const workbookId = workbook.getUnitId();
+                        const worksheetId = this._univerInstanceService
+                            .getCurrentUniverSheetInstance()
+                            .getActiveSheet()
+                            .getSheetId();
+                        const model = this._numfmtService.getModel(workbookId, worksheetId);
+                        const disposableMap: Map<string, IDisposable> = new Map();
+                        const register = (commandInfo: EffectRefRangeParams, row: number, col: number) => {
+                            const targetRange = {
+                                startRow: row,
+                                startColumn: col,
+                                endRow: row,
+                                endColumn: col,
+                            };
+                            let operators = [];
+                            switch (commandInfo.id) {
+                                case EffectRefRangId.DeleteRangeMoveLeftCommandId: {
+                                    operators = handleDeleteRangeMoveLeft(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.DeleteRangeMoveUpCommandId: {
+                                    operators = handleDeleteRangeMoveUp(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.InsertColCommandId: {
+                                    operators = handleInsertCol(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.InsertRangeMoveDownCommandId: {
+                                    operators = handleInsertRangeMoveDown(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.InsertRangeMoveRightCommandId: {
+                                    operators = handleInsertRangeMoveRight(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.InsertRowCommandId: {
+                                    operators = handleInsertRow(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.MoveRangeCommandId: {
+                                    operators = handleMoveRange(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.RemoveColCommandId: {
+                                    operators = handleIRemoveCol(commandInfo, targetRange);
+                                    break;
+                                }
+                                case EffectRefRangId.RemoveRowCommandId: {
+                                    operators = handleIRemoveRow(commandInfo, targetRange);
+                                    break;
+                                }
+                            }
+                            const result = runRefRangeMutations(operators, targetRange);
+                            if (!result) {
+                                // 删除
+                                const redoParams: SetNumfmtMutationParams = {
+                                    workbookId,
+                                    worksheetId,
+                                    values: [{ row, col }],
+                                };
+                                const undoParams = factorySetNumfmtUndoMutation(this._injector, redoParams);
+                                return {
+                                    redos: [
+                                        {
+                                            id: SetNumfmtMutation.id,
+                                            params: redoParams,
+                                        },
+                                    ],
+                                    undos: [
+                                        {
+                                            id: SetNumfmtMutation.id,
+                                            params: undoParams,
+                                        },
+                                    ],
+                                };
+                            }
+                            const numfmtValue = this._numfmtService.getValue(workbookId, worksheetId, row, col);
+
+                            if (numfmtValue) {
+                                const redoParams: SetNumfmtMutationParams = {
+                                    workbookId,
+                                    worksheetId,
+                                    values: [
+                                        { row, col },
+                                        { ...numfmtValue, row: result.startRow, col: result.startColumn },
+                                    ],
+                                };
+                                const undoParams = factorySetNumfmtUndoMutation(this._injector, redoParams);
+                                return {
+                                    redos: [
+                                        {
+                                            id: SetNumfmtMutation.id,
+                                            params: redoParams,
+                                        },
+                                    ],
+                                    undos: [
+                                        {
+                                            id: SetNumfmtMutation.id,
+                                            params: undoParams,
+                                        },
+                                    ],
+                                };
+                            }
+                            return { redos: [], undos: [] };
+                        };
+                        if (model) {
+                            model.forValue((row, col) => {
+                                const targetRange = {
+                                    startRow: row,
+                                    startColumn: col,
+                                    endRow: row,
+                                    endColumn: col,
+                                };
+                                const disposable = this._refRangeService.registerRefRange(targetRange, (commandInfo) =>
+                                    register(commandInfo, row, col)
+                                );
+                                disposableMap.set(`${row}_${col}`, disposable);
+                                disposableCollection.add(disposable);
+                            });
+                        }
+
+                        // on change
+                        disposableCollection.add(
+                            toDisposable(
+                                new Observable<SetNumfmtMutationParams>((subscribe) => {
+                                    disposableCollection.add(
+                                        this._commandService.onCommandExecuted((commandInfo) => {
+                                            if (commandInfo.id === SetNumfmtMutation.id) {
+                                                subscribe.next(commandInfo.params as SetNumfmtMutationParams);
+                                            }
+                                        })
+                                    );
+                                })
+                                    .pipe(
+                                        filter(
+                                            (commandInfo) =>
+                                                commandInfo.workbookId === workbookId &&
+                                                commandInfo.worksheetId === worksheetId
+                                        ),
+                                        map((commandInfo) => commandInfo.values),
+                                        bufferTime(300), // updating the listener moves to the next queue
+                                        map((values) =>
+                                            values.reduce((pre, cur) => {
+                                                pre.push(...cur);
+                                                return pre;
+                                            }, [])
+                                        ),
+                                        filter((values) => !!values.length)
+                                    )
+                                    .subscribe((values) => {
+                                        values.forEach((value) => {
+                                            const { row, col } = value;
+                                            const key = `${row}_${col}`;
+                                            const disposable = disposableMap.get(key);
+                                            disposableMap.delete(key);
+                                            if (!value.pattern) {
+                                                disposable && disposable.dispose();
+                                            } else {
+                                                const targetRange = {
+                                                    startRow: row,
+                                                    startColumn: col,
+                                                    endRow: row,
+                                                    endColumn: col,
+                                                };
+                                                const disposable = this._refRangeService.registerRefRange(
+                                                    targetRange,
+                                                    (commandInfo) => register(commandInfo, row, col)
+                                                );
+                                                disposableMap.set(key, disposable);
+                                                disposableCollection.add(disposable);
+                                            }
+                                        });
+                                    })
+                            )
+                        );
+                    })
+            )
+        );
+    }
+}
+
+const filterAtr = (obj: Record<any, any>, filterKey: string[]) => {
+    const keys = Object.keys(obj).filter((key) => !filterKey.includes(key));
+    return keys.reduce(
+        (pre, cur) => {
+            pre[cur] = obj[cur];
+            return pre;
+        },
+        {} as Record<any, any>
+    );
+};
+
+function isNumeric(str: string) {
+    return /^-?\d+(\.\d+)?$/.test(str);
+}
